Fix HandleNextCounterCell (TIME2 is not incrementing on TIME1 overflow)
The reason is that State->HighestPriorityCounter may have been modified
by one of the counter instructions (e.g. CounterPINC with an overflow of
TIME1). If that is the case, then the increment-by-one skips over the
counter, and the count is permanently missed.

This costs just one more iteration if State->HighestPriorityCounter is
unchanged by the call to HandleNextCounterCell.

See virtualagc/virtualagc#1060